Report total packages load and activation times
This reports a sum of how long it took for Atom to load all packages
and how long to activate them.

The point is that this can be used to track if / how much of a problem
package load times are for users, and how these times develop over time.

I have tested this by adding console.log() output before the
sendTiming() lines and verified that the reported numbers are close
to what TimeCop says.